LPS-199482 Transform class JAXRSLifecycle to class util, and apply to usages

# breaking

## What modules/apps/portal-remote/portal-remote-jaxrs-whiteboard/src/main/java/com/liferay/portal/remote/jaxrs/whiteboard/lifecycle/JAXRSLifecycle.java

modules/apps/portal-remote/portal-remote-jaxrs-whiteboard/src/main/java/com/liferay/portal/remote/jaxrs/whiteboard/lifecycle/JAXRSLifecycle.java Was removed

## Why

This class was turned into a static class because it had a component with its own class name as a service type

## Alternatives

Use modules/apps/portal-remote/portal-remote-jaxrs-whiteboard/src/main/java/com/liferay/portal/remote/jaxrs/whiteboard/util/JAXRSLifecycleUtil.java directly and call the desired method

----
