+++ modules/apps/oauth2-provider/oauth2-provider-scope-impl/src/test/java/com/liferay/oauth2/provider/scope/internal/liferay/ScopeLocatorImplTest.java
@@ -20,7 +20,6 @@ import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
 import com.liferay.portal.kernel.module.util.SystemBundleUtil;
 import com.liferay.portal.kernel.test.ReflectionTestUtil;
 import com.liferay.portal.kernel.util.PropsUtil;
-import com.liferay.portal.remote.jaxrs.whiteboard.lifecycle.JAXRSLifecycle;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
 import com.liferay.portal.util.PropsImpl;
 
@@ -449,16 +448,6 @@ public class ScopeLocatorImplTest {
 					registrator -> {
 					});
 			}
-
-			ReflectionTestUtil.setFieldValue(
-				_scopeLocatorImpl, "_jaxrsLifecycle",
-				new JAXRSLifecycle() {
-
-					@Override
-					public void ensureReady() {
-					}
-
-				});
 
 			return _scopeLocatorImpl;
 		}
