+++ modules/apps/portal-remote/portal-remote-jaxrs-whiteboard/bnd.bnd
@@ -1,8 +1,8 @@
 Bundle-Name: Liferay Portal Remote JAXRS Whiteboard
 Bundle-SymbolicName: com.liferay.portal.remote.jaxrs.whiteboard
-Bundle-Version: 12.0.2
+Bundle-Version: 13.0.0
 Export-Package:\
-	com.liferay.portal.remote.jaxrs.whiteboard.lifecycle,\
+	com.liferay.portal.remote.jaxrs.whiteboard.util,\
 	\
 	javax.wsdl;version="1.2",\
 	javax.wsdl.extensions;version="1.2",\
