+++ modules/apps/portal-remote/portal-remote-jaxrs-whiteboard/src/main/resources/com/liferay/portal/remote/jaxrs/whiteboard/util/packageinfo
